Allow the user/pw or org/key to be Deferred.
This introduces conditional logic to detect if sensitive/secret values
may be passed in as Deferred type, and if so to process them entirely
through a chain of Deferred arguments, to address issues with
interpolating Deferred into strings on the compiler.

Unfortunately, there is no way at the moment to pass a parameter of type
Deferred into puppet-rspec, so while this has been tested and verified
locally, there is not currently a unit test for this new functionality.